Move beta groups to testflight folder
