+++ src/codemagic/apple/app_store_connect/provisioning/__init__.py
@@ -1,4 +1,3 @@
-from .beta_groups import BetaGroups
 from .bundle_id_capabilities import BundleIdCapabilities
 from .bundle_ids import BundleIds
 from .devices import Devices
