+++ src/codemagic/apple/app_store_connect/testflight/__init__.py
@@ -1,0 +1,1 @@
+from .beta_groups import BetaGroups
