+++ src/codemagic/apple/app_store_connect/testflight/beta_groups.py
